requirements.txt
pytse-client==0.18.0
scipy==1.10.1
networkx==3.1
matplotlib==3.7.4
arabic-reshaper==3.0.0
python-bidi==0.4.2
pyomo==6.7.0